Split RSATwoPowerMod
