+++ systemc/CMakeLists.txt
@@ -10,10 +10,10 @@ set(SYSTEMC_LIBRARY rsa systemc)
 
 target_link_libraries(RSA256 ${SYSTEMC_LIBRARY})
 
-add_executable(RSATwoPowerMod RSATwoPowerMod.cpp)
 
-target_link_libraries(RSATwoPowerMod ${SYSTEMC_LIBRARY})
-
+add_library(RSATwoPowerMod RSATwoPowerMod.cpp)
+add_executable(RSATwoPowerMod_test RSATwoPowerMod_test.cpp)
+target_link_libraries(RSATwoPowerMod_test ${SYSTEMC_LIBRARY} RSATwoPowerMod)
 
 add_library(RSAMontgomery RSAMontgomery.cpp)
 add_executable(RSAMontgomery_test RSAMontgomery_test.cpp)
